Refactor the code moving Bookmark, Card, CardList, Form, Header and Data to independent files.
